MAINT: rename to bidsviewmodel
