+++ applications/mne_analyze/libs/anShared/Management/analyzedata.cpp
@@ -64,7 +64,7 @@ using namespace ANSHAREDLIB;
 
 AnalyzeData::AnalyzeData(QObject *pParent)
 : QObject(pParent)
-, m_pData(new AnalyzeDataModel(this))
+, m_pData(new BidsViewModel(this))
 {
 
 }
